Major: now refreshing the store, when close to overflow
When store is close to its maximum value at least in one byte, it is set to zero and bytes in it are summed and added to ans. Moreover, last third part is also completed. Now just debugging left and also implementing palignr instruction, which requires further analyzing. Nevertheless, main part of work is done